+++ src/rez/tests/test_packages.py
@@ -3,9 +3,11 @@ test package iteration, serialization etc
 """
 from rez.packages import iter_package_families, iter_packages, get_package, \
     create_package, get_developer_package, get_variant_from_uri, \
-    get_package_from_uri
+    get_package_from_uri, get_package_from_repository
 from rez.package_py_utils import expand_requirement
 from rez.package_resources import package_release_keys
+from rez.package_move import move_package
+from rez.package_remove import remove_package, remove_packages_ignored_since
 from rez.package_repository import package_repository_manager
 from rez.tests.util import TestBase, TempdirMixin
 from rez.utils.formatting import PackageRequest
@@ -444,16 +446,28 @@ class TestPackages(TestBase, TempdirMixin):
         pkg_version = Version("2")
 
         # copy packages to a temp repo
-        repo_path = os.path.join(self.root, "tmp_packages")
+        repo_path = os.path.join(self.root, "tmp1_packages")
         shutil.copytree(self.solver_packages_path, repo_path)
 
         # check that a known package exists
-        pkg = get_package(pkg_name, pkg_version, paths=[repo_path])
+        pkg = get_package_from_repository(pkg_name, pkg_version, repo_path)
         self.assertNotEqual(pkg, None)
 
         repo = pkg.repository
 
-        # ignore the package
+        # ignore a pkg that doesn't exist
+        i = repo.ignore_package("i_dont_exist", Version("1"))
+        self.assertEqual(i, -1)
+
+        # ignore a pkg that doesn't exist, but allow it
+        i = repo.ignore_package("i_dont_exist", Version("1"), allow_missing=True)
+        self.assertEqual(i, 1)
+
+        # unignore a pkg that doesn't exist
+        i = repo.unignore_package("i_dont_exist", Version("1"))
+        self.assertEqual(i, -1)
+
+        # ignore an existing package
         i = repo.ignore_package(pkg_name, pkg_version)
         self.assertEqual(i, 1)
 
@@ -462,8 +476,7 @@ class TestPackages(TestBase, TempdirMixin):
         self.assertEqual(i, 0)
 
         # verify that we cannot see it
-        package_repository_manager.clear_caches()
-        pkg = get_package(pkg_name, pkg_version, paths=[repo_path])
+        pkg = get_package_from_repository(pkg_name, pkg_version, repo_path)
         self.assertEqual(pkg, None)
 
         # unignore it
@@ -475,9 +488,83 @@ class TestPackages(TestBase, TempdirMixin):
         self.assertEqual(i, 0)
 
         # verify that we can see it again
-        package_repository_manager.clear_caches()
-        pkg = get_package(pkg_name, pkg_version, paths=[repo_path])
+        pkg = get_package_from_repository(pkg_name, pkg_version, repo_path)
         self.assertNotEqual(pkg, None)
+
+    def test_package_move(self):
+        """Test package move."""
+        pkg_name = "pydad"
+        pkg_version = Version("2")
+
+        # copy packages to a temp repo
+        repo_path = os.path.join(self.root, "tmp2_packages")
+        shutil.copytree(self.solver_packages_path, repo_path)
+
+        # create an empty temp repo
+        dest_repo_path = os.path.join(self.root, "tmp3_packages")
+        os.mkdir(dest_repo_path)
+
+        # verify that source pkg exists
+        src_pkg = get_package_from_repository(pkg_name, pkg_version, repo_path)
+        self.assertNotEqual(src_pkg, None)
+
+        # move it to dest repo
+        move_package(src_pkg, dest_repo_path)
+
+        # verify it exists in dest repo
+        dest_pkg = get_package_from_repository(pkg_name, pkg_version, dest_repo_path)
+        self.assertNotEqual(dest_pkg, None)
+
+        # verify it is not visible in source repo
+        src_pkg = get_package_from_repository(pkg_name, pkg_version, repo_path)
+        self.assertEqual(src_pkg, None)
+
+    def test_package_remove(self):
+        """Test package remove."""
+        pkg_name = "pydad"
+        pkg_version = Version("2")
+
+        # copy packages to a temp repo
+        repo_path = os.path.join(self.root, "tmp4_packages")
+        shutil.copytree(self.solver_packages_path, repo_path)
+
+        # verify that source pkg exists
+        src_pkg = get_package_from_repository(pkg_name, pkg_version, repo_path)
+        self.assertNotEqual(src_pkg, None)
+
+        # remove it
+        was_removed = remove_package(pkg_name, pkg_version, repo_path)
+        self.assertTrue(was_removed)
+
+        # verify that source pkg no longer exists (and isn't simply ignored)
+        repo = package_repository_manager.get_repository(repo_path)
+        i = repo.unignore_package(pkg_name, pkg_version)
+        self.assertEqual(i, -1)
+
+    def test_remove_packages_ignored_since(self):
+        pkg_name = "pydad"
+        pkg_version = Version("2")
+
+        # copy packages to a temp repo
+        repo_path = os.path.join(self.root, "tmp5_packages")
+        shutil.copytree(self.solver_packages_path, repo_path)
+
+        # verify that source pkg exists
+        src_pkg = get_package_from_repository(pkg_name, pkg_version, repo_path)
+        self.assertNotEqual(src_pkg, None)
+
+        # ignore it
+        repo = package_repository_manager.get_repository(repo_path)
+        i = repo.ignore_package(pkg_name, pkg_version)
+        self.assertEqual(i, 1)
+
+        # remove all ignored packages
+        num_removed = remove_packages_ignored_since(days=0, paths=[repo_path])
+        self.assertEqual(num_removed, 1)
+
+        # verify that source pkg no longer exists (and isn't simply ignored)
+        i = repo.unignore_package(pkg_name, pkg_version)
+        self.assertEqual(i, -1)
 
 
 class TestMemoryPackages(TestBase):